+++ appservice/package.json
@@ -1,11 +1,7 @@
 {
     "name": "vscode-azureappservice",
     "author": "Microsoft Corporation",
-<<<<<<< HEAD
-    "version": "0.29.3",
-=======
     "version": "0.30.0",
->>>>>>> 32e2625cc1df1c9e4f231c07bc95edb27a304c40
     "description": "Common tools for developing Azure App Service extensions for VS Code",
     "tags": [
         "azure",
