+++ appservice/src/deploy/runPreDeployTask.ts
@@ -9,9 +9,6 @@ import { ext } from '../extensionVariables';
 import { localize } from '../localize';
 import { ScmType } from '../ScmType';
 import { isPathEqual, isSubpath } from '../utils/pathUtils';
-import { SiteClient } from '../SiteClient';
-import { StringDictionary } from 'azure-arm-website/lib/models';
-import { formatDeployLog } from './formatDeployLog';
 
 export async function runPreDeployTask(actionContext: IActionContext, deployFsPath: string, scmType: string | undefined, extensionPrefix: string): Promise<void> {
     const preDeployResult: IPreDeployTaskResult = await tryRunPreDeployTask(actionContext, deployFsPath, scmType, extensionPrefix);
@@ -97,22 +94,5 @@ export async function handleFailedPreDeployTask(actionContext: IActionContext, p
     } else {
         actionContext.properties.preDeployTaskResponse = 'cancel';
         throw new UserCancelledError();
-    }
-}
-
-// prior to git deploying, these settings must be deleted or it will fail
-export async function verifyNoRunFromPackageSetting(client: SiteClient): Promise<void> {
-    let updateSettings: boolean = false;
-    const runFromPackageSettings: string[] = ['WEBSITE_RUN_FROM_PACKAGE', 'WEBSITE_RUN_FROM_ZIP'];
-    const applicationSettings: StringDictionary = await client.listApplicationSettings();
-    for (const settingName of runFromPackageSettings) {
-        if (applicationSettings.properties && applicationSettings.properties[settingName]) {
-            delete applicationSettings.properties[settingName];
-            ext.outputChannel.appendLine(formatDeployLog(client, localize('deletingSetting', 'Deleting setting "{0}"...', settingName)));
-            updateSettings = true;
-        }
-    }
-    if (updateSettings) {
-        await client.updateApplicationSettings(applicationSettings);
     }
 }
